Moved several files to new folders.  Gave IContractProvider indxer property.
