+++ Traction/IContractProvider.cs
@@ -1,8 +1,12 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 
 namespace Traction {
 
     interface IContractProvider {
+
         IEnumerable<Contract> Contracts { get; }
+
+        Contract this[Type attributeType] { get; }
     }
 }
