+++ Traction/Rewriters/DiagnosticChecker.cs
